--- fpsim/locations/kenya/data/splines_25.csv
@@ -1,36 +1,0 @@
-"knot_1","knot_2","age"
-0,0,15
-0.0540909163346256,-0.0316577582570118,16
-0.107710562569228,-0.0627349950963628,17
-0.160387668603784,-0.0926511891003925,18
-0.21165096433827,-0.12082581885144,19
-0.261029179672662,-0.146678362931845,20
-0.308051044506938,-0.169628299923946,21
-0.352245288741075,-0.189095108410082,22
-0.393140642275048,-0.204498266972594,23
-0.430265835008834,-0.21525725419382,24
-0.463149596842412,-0.220791548656099,25
-0.491431929782705,-0.220657696498719,26
-0.515197924264443,-0.214960514088757,27
-0.5346439428293,-0.203941885350239,28
-0.549966348018953,-0.187843694207189,29
-0.561361502375078,-0.166907824583633,30
-0.569025768439353,-0.141376160403596,31
-0.573155508753453,-0.111490585591104,32
-0.573947085859054,-0.0774929840701823,33
-0.571596862297833,-0.0396252397648555,34
-0.566301200611467,0.00187076340085084,35
-0.558256463341631,0.0467531415029113,36
-0.547659013030003,0.0947800106173008,37
-0.534705212218258,0.145709486819994,38
-0.519591423448072,0.199299686186965,39
-0.502514009261123,0.255308724794189,40
-0.483669332199087,0.313494718717641,41
-0.46325375480364,0.373615784033296,42
-0.441463639616458,0.435430036817127,43
-0.418495349179217,0.498695593145111,44
-0.394545246033595,0.56317056909322,45
-0.369809692721268,0.628613080737431,46
-0.344485051783911,0.694781244153718,47
-0.318767685763202,0.761433175418056,48
-0.292853957200816,0.828326990606418,49
