--- fpsim/locations/kenya/data/splines_25_40.csv
@@ -1,36 +1,0 @@
-"knot_1","knot_2","knot_3","age"
-0,0,0,15
--0.0175912222339768,0.0649325207402679,-0.0472236514474676,16
--0.0342211093077761,0.128928381187336,-0.0937660954089716,17
--0.0489283260612204,0.191050921048004,-0.138946124398548,18
--0.060751537334132,0.250363480029072,-0.182082530930234,19
--0.0687294079663336,0.305929397837341,-0.222494107518066,20
--0.0719006027976475,0.35681201417961,-0.25949964667608,21
--0.0693037866678963,0.402074668762679,-0.292417940918312,22
--0.0599776244169024,0.440780701293348,-0.320567782758799,23
--0.0429607808844883,0.471993451478418,-0.343267964711577,24
--0.0172919209104765,0.494776259024688,-0.359837279290682,25
-0.0175377405332322,0.508524653629419,-0.369751936646312,26
-0.0602267879464232,0.513962924951713,-0.3731178174733,27
-0.109021255696804,0.512147552641131,-0.370198220102641,28
-0.162167178152081,0.504135016347235,-0.361256442865329,29
-0.217910589679962,0.490981795719588,-0.34655578409236,30
-0.274497524648154,0.47374437040775,-0.326359542114727,31
-0.330174017424364,0.453479220061284,-0.300931015263425,32
-0.383186102376299,0.431242824329752,-0.270533501869449,33
-0.431779813871666,0.408091662862716,-0.235430300263793,34
-0.474201186278172,0.385082215309737,-0.195884708777452,35
-0.508696253963526,0.363270961320378,-0.152160025741419,36
-0.533511051295432,0.3437143805442,-0.104519549486691,37
-0.5468916126416,0.327468952630765,-0.0532265783442602,38
-0.547083972369735,0.315591157229636,0.00145558935487777,39
-0.532334164847545,0.309137473990373,0.0592636552797285,40
-0.501489739694094,0.308868291661475,0.119925193045302,41
-0.455804307533872,0.314359635387181,0.183131264050625,42
-0.397132994242724,0.324891439410664,0.248563801640729,43
-0.327330925696498,0.3397436379751,0.315904739160645,44
-0.24825322777104,0.358196165323664,0.384836009955405,45
-0.161755026342196,0.379528955699529,0.45503954737004,46
-0.069691447285813,0.40302194334587,0.52619728474958,47
--0.0260823835222631,0.427955062505863,0.597991155439058,48
--0.123711340206186,0.453608247422681,0.670103092783505,49
