+++ tests/test_sim.py
@@ -11,9 +11,9 @@ par_kwargs = dict(n_agents=500, start_year=2000, end_year=2010, seed=1, verbose=
 serial = 1  # Whether to run in serial (for debugging)
 
 
-def test_simple():
+def test_simple(location='kenya'):
     sc.heading('Test simplest possible FPsim run')
-    sim = fp.Sim()
+    sim = fp.Sim(location=location)
     sim.run()
     return sim
 
@@ -54,7 +54,7 @@ def test_mid_choice():
 
 if __name__ == '__main__':
 
-    # s0 = test_simple()
-    sims1 = test_simple_choice()
+    s0 = test_simple('ethiopia')
+    # sims1 = test_simple_choice()
     # sims2 = test_mid_choice()
     print('Done.')
